Move members view to new directory
